User Roles

+++ app/controllers/wikis_controller.rb
@@ -1,6 +1,11 @@
 class WikisController < ApplicationController
+
+  def user_wikis
+    @wikis = current_user.own_wikis
+  end
+
   def index
-    @wikis = Wiki.all
+    @wikis = policy_scope(Wiki)
   end
 
   def show
@@ -12,9 +17,8 @@ class WikisController < ApplicationController
   end
 
   def create
-    @wiki = Wiki.new
-    @wiki.title = params[:wiki][:title]
-    @wiki.body = params[:wiki][:body]
+    @wiki = current_user.wikis.build(params.require(:wiki).permit(:title, :body, :private))
+
     if @wiki.save
       flash[:notice] = "Wiki was saved successfully."
       redirect_to @wiki
@@ -26,16 +30,16 @@ class WikisController < ApplicationController
 
   def edit
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def update
     @wiki = Wiki.find(params[:id])
-    @wiki.title = params[:wiki][:title]
-    @wiki.body = params[:wiki][:body]
+    authorize @wiki
 
-    if @wiki.save
+    if @wiki.update_attributes(params.require(:wiki).permit(:title, :body, :private))
       flash[:notice] = "Wiki was updated successfully."
-      redirect_to @wiki
+      redirect_to edit_wiki_path(@wiki)
     else
       flash.now[:alert] = "There was an error saving the wiki. Please try again."
       render :edit
@@ -44,10 +48,12 @@ class WikisController < ApplicationController
 
   def destroy
     @wiki = Wiki.find(params[:id])
+    title = @wiki.title
+    authorize @wiki
 
     if @wiki.destroy
-      flash[:notice] = "\"#{@wiki.title}\" was deleted successfully."
-      redirect_to wikis_path
+      flash[:notice] = "\"#{title}\" was deleted successfully."
+      redirect_to @wiki
    else
       flash.now[:alert] = "There was an error deleting the wiki."
       render :show
